Reorganize error classes
