+++ packages/core/src/index.ts
@@ -1,8 +1,6 @@
 export * from "./nack-errors/NackCode.js";
 export * from "./ExpansionHub.js";
-export * from "./serial-errors.js";
 export * from "./RevHub.js";
-export * from "./RevHubError.js";
 export * from "./RevHubType.js";
 export * from "./led-pattern.js";
 export * from "./BulkInputData.js";
@@ -23,17 +21,18 @@ export * from "./Rgb.js";
 export * from "./MotorMode.js";
 export * from "./VerbosityLevel.js";
 export * from "./Version.js";
-export * from "./serial-errors.js";
+export * from "./general-errors/NoExpansionHubWithAddressError.js";
+export * from "./general-errors/RevHubError.js";
+export * from "./general-errors/serial-errors.js";
+export * from "./general-errors/set-prototype-of.js";
+export * from "./general-errors/TimeoutError.js";
 export * from "./nack-errors/NackError.js";
 export * from "./nack-errors/BatteryTooLowError.js";
 export * from "./nack-errors/diagnostic-errors.js";
 export * from "./nack-errors/digital-channel-errors.js";
-export * from "./nack-errors/GeneralSerialError.js";
 export * from "./nack-errors/i2c-errors.js";
 export * from "./nack-errors/motor-errors.js";
 export * from "./nack-errors/nack-code-to-error.js";
-export * from "./nack-errors/NoExpansionHubWithAddressError.js";
 export * from "./nack-errors/ParameterOutOfRangeError.js";
 export * from "./nack-errors/servo-errors.js";
-export * from "./nack-errors/TimeoutError.js";
 export * from "./nack-errors/UnrecognizedNackError.js";
