--- packages/core/src/nack-errors/GeneralSerialError.ts
@@ -1,14 +1,0 @@
-import { setPrototypeOf } from "./NackError.js";
-
-/**
- * Indicates an error in the Serial connection.
- */
-export class GeneralSerialError extends Error {
-    serialNumber: string;
-    constructor(serialNumber: string) {
-        super(`Serial Port Error for ${serialNumber}`);
-
-        this.serialNumber = serialNumber;
-        setPrototypeOf(this, GeneralSerialError.prototype);
-    }
-}
