+++ packages/core/src/nack-errors/UnrecognizedNackError.ts
@@ -1,4 +1,5 @@
-import { NackError, setPrototypeOf } from "./NackError.js";
+import { setPrototypeOf } from "../general-errors/set-prototype-of.js";
+import { NackError } from "./NackError.js";
 
 export class UnrecognizedNackError extends NackError {
     constructor(nackCode: number) {
